Tabs: Combine and cleanup toolbar controls
