+++ packages/block-library/src/tab-list/edit.js
@@ -21,8 +21,7 @@ import { useCallback, useEffect, useRef } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import AddTabToolbarControl from '../tab-panel/add-tab-toolbar-control';
-import RemoveTabToolbarControl from '../tab-panel/remove-tab-toolbar-control';
+import TabToolbarControls from '../tab-panel/tab-toolbar-controls';
 
 const EMPTY_ARRAY = [];
 
@@ -138,8 +137,7 @@ function Edit( {
 
 	return (
 		<>
-			<AddTabToolbarControl tabsClientId={ tabsClientId } />
-			<RemoveTabToolbarControl tabsClientId={ tabsClientId } />
+			<TabToolbarControls tabsClientId={ tabsClientId } />
 			<div { ...blockProps }>
 				{ tabsList.map( ( tab, index ) => {
 					const isActive = index === effectiveActiveIndex;
