+++ packages/block-library/src/tab-panel/controls.js
@@ -16,8 +16,7 @@ import { useDispatch } from '@wordpress/data';
 /**
  * Internal dependencies
  */
-import AddTabToolbarControl from './add-tab-toolbar-control';
-import RemoveTabToolbarControl from './remove-tab-toolbar-control';
+import TabToolbarControls from './tab-toolbar-controls';
 import { useToolsPanelDropdownMenuProps } from '../utils/hooks';
 
 export default function Controls( { tabsClientId, blockIndex, isDefaultTab } ) {
@@ -26,8 +25,7 @@ export default function Controls( { tabsClientId, blockIndex, isDefaultTab } ) {
 
 	return (
 		<>
-			<AddTabToolbarControl tabsClientId={ tabsClientId } />
-			<RemoveTabToolbarControl tabsClientId={ tabsClientId } />
+			<TabToolbarControls tabsClientId={ tabsClientId } />
 			<InspectorControls>
 				<ToolsPanel
 					label={ __( 'Settings' ) }
