+++ packages/block-library/src/tab-panel/tab-toolbar-controls.js
@@ -1,25 +1,27 @@
 /**
  * WordPress dependencies
  */
+import { __ } from '@wordpress/i18n';
+import { createBlock } from '@wordpress/blocks';
 import {
 	BlockControls,
 	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { ToolbarGroup, ToolbarButton } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
 import { useDispatch, useSelect } from '@wordpress/data';
 
 /**
- * "Remove Tab" button in the block toolbar for the tabs block.
- * Removes the currently active core/tab-panel. The tab-list items
- * attribute is kept in sync automatically via useTabListItemsSync.
+ * "Add tab" and "Remove tab" buttons in the block toolbar for the tabs block.
+ * Inserts or removes a core/tab-panel. The tab-list items attribute is kept
+ * in sync automatically via useTabListItemsSync.
  *
  * @param {Object} props
  * @param {string} props.tabsClientId The client ID of the parent tabs block.
  * @return {React.JSX.Element} The toolbar control element.
  */
-export default function RemoveTabToolbarControl( { tabsClientId } ) {
+export default function TabToolbarControls( { tabsClientId } ) {
 	const {
+		insertBlock,
 		removeBlock,
 		updateBlockAttributes,
 		selectBlock,
@@ -27,6 +29,7 @@ export default function RemoveTabToolbarControl( { tabsClientId } ) {
 	} = useDispatch( blockEditorStore );
 
 	const {
+		tabPanelsClientId,
 		activeTabPanelClientId,
 		tabCount,
 		editorActiveTabIndex,
@@ -35,6 +38,7 @@ export default function RemoveTabToolbarControl( { tabsClientId } ) {
 		( select ) => {
 			if ( ! tabsClientId ) {
 				return {
+					tabPanelsClientId: null,
 					activeTabPanelClientId: null,
 					tabCount: 0,
 					editorActiveTabIndex: 0,
@@ -59,6 +63,7 @@ export default function RemoveTabToolbarControl( { tabsClientId } ) {
 			const activeTabPanel = tabPanelBlocks[ activeIndex ];
 
 			return {
+				tabPanelsClientId: tabPanels?.clientId || null,
 				activeTabPanelClientId: activeTabPanel?.clientId || null,
 				tabCount: tabPanelBlocks.length,
 				editorActiveTabIndex: activeIndex,
@@ -67,6 +72,29 @@ export default function RemoveTabToolbarControl( { tabsClientId } ) {
 		},
 		[ tabsClientId ]
 	);
+
+	const addTab = () => {
+		if ( ! tabPanelsClientId ) {
+			return;
+		}
+
+		const newTabPanelBlock = createBlock( 'core/tab-panel', {
+			label: __( 'Tab' ),
+		} );
+		insertBlock( newTabPanelBlock, undefined, tabPanelsClientId, false );
+
+		// Switch editor active tab to the new tab.
+		const newIndex = tabCount;
+		__unstableMarkNextChangeAsNotPersistent();
+		updateBlockAttributes( tabsClientId, {
+			editorActiveTabIndex: newIndex,
+		} );
+
+		// Select the tab-list block so focus stays in the menu area.
+		if ( tabListClientId ) {
+			selectBlock( tabListClientId );
+		}
+	};
 
 	const removeTab = () => {
 		if ( ! activeTabPanelClientId || tabCount <= 1 ) {
@@ -93,16 +121,21 @@ export default function RemoveTabToolbarControl( { tabsClientId } ) {
 		}
 	};
 
-	const isDisabled = tabCount <= 1 || ! activeTabPanelClientId;
+	const isRemoveDisabled = tabCount <= 1 || ! activeTabPanelClientId;
 
 	return (
 		<BlockControls group="other">
 			<ToolbarGroup>
 				<ToolbarButton
 					className="components-toolbar__control"
+					onClick={ addTab }
+					text={ __( 'Add tab' ) }
+				/>
+				<ToolbarButton
+					className="components-toolbar__control"
 					onClick={ removeTab }
 					text={ __( 'Remove tab' ) }
-					disabled={ isDisabled }
+					disabled={ isRemoveDisabled }
 				/>
 			</ToolbarGroup>
 		</BlockControls>
